Enable the `Generic.CodeAnalysis.UnusedFunctionParameter` sniff.
The reason why the sniff was disabled up to now was that WP uses callbacks a lot and the hooked in functions may not use all parameters passed to them.

This concern was previously raised in https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/pull/382#discussion_r29981655

To this end, I've send in a PR upstream to address these concerns, which has subsequently been merged.
The adjusted version of the sniff will be released in PHPCS 2.7.2.